refactor(client/useAuth): implement logic in the same file
