+++ client/src/customHooks/useAuth.jsx
@@ -1,9 +1,5 @@
-import { Utils } from "stellar-sdk";
-import {
-  readChallenge,
-  signChallenge,
-  sendChallenge,
-} from "./functionsAuthentication";
+import { Utils, Keypair, Transaction } from "stellar-sdk";
+import axios from "axios";
 
 export default async function useAuth({
   authEndpoint,
@@ -12,7 +8,7 @@ export default async function useAuth({
   secretKey,
 }) {
   try {
-    const tx = await readChallenge({
+    const tx = await startChallenge({
       authEndpoint,
       serverPublicKey,
       publicKey,
@@ -20,12 +16,69 @@ export default async function useAuth({
 
     const transaction = await signChallenge({ tx, secretKey });
 
-    let token;
-    if (Utils.verifyTxSignedBy(transaction, publicKey))
-      token = await sendChallenge({ authEndpoint, transaction });
-
-    return token;
+    if (
+      transaction &&
+      Utils.verifyTxSignedBy(transaction, publicKey) &&
+      Utils.validateTimebounds(transaction)
+    ) {
+      let token = await sendChallenge({ authEndpoint, transaction });
+      return token;
+    }
+    return false;
   } catch (error) {
     return error;
   }
 }
+
+const startChallenge = async ({ authEndpoint, serverPublicKey, publicKey }) => {
+  const params = { account: publicKey };
+
+  const authURL = new URL(authEndpoint);
+
+  Object.entries(params).forEach(([key, value]) => {
+    authURL.searchParams.append(key, value);
+  });
+
+  const result = await axios.get(authURL.toString());
+
+  if (!result.transaction) {
+    throw new Error("The response didn’t contain a transaction");
+  }
+
+  const { tx } = Utils.readChallengeTx(
+    result.transaction,
+    serverPublicKey,
+    result.network_passphrase,
+    authURL.host
+  );
+
+  return tx;
+};
+
+const signChallenge = ({ tx, secretKey }) => {
+  const envelope = tx.result.transaction.toEnvelope().toXDR("base64");
+  const transaction = new Transaction(envelope, tx.result.network_passphrase);
+  transaction.sign(Keypair.fromSecret(secretKey));
+
+  return transaction;
+};
+
+const sendChallenge = async ({ authEndpoint, transaction }) => {
+  const params = {
+    transaction: transaction.toEnvelope().toXDR("base64"),
+  };
+
+  const urlParams = new URLSearchParams(params);
+
+  const result = await axios({
+    method: "POST",
+    url: authEndpoint,
+    data: urlParams.toString(),
+    headers: { "content-type": "application/x-www-form-urlencoded" },
+  });
+  if (!result.token) {
+    return new Error("No token returned from `/auth`");
+  }
+
+  return result.token;
+};
